+++ progs/cicada_extract_scfg_impl.hpp
@@ -1114,9 +1114,9 @@ struct ExtractSCFG
     if (inverse) {
       alignment_type::const_iterator aiter_end = alignment.end();
       for (alignment_type::const_iterator aiter = alignment.begin(); aiter != aiter_end; ++ aiter) {
-	if (aiter->target >= alignment_source_target.size())
+	if (aiter->target >= static_cast<int>(alignment_source_target.size()))
 	  throw std::runtime_error("invalid alignment");
-	if (aiter->source >= alignment_target_source.size())
+	if (aiter->source >= static_cast<int>(alignment_target_source.size()))
 	  throw std::runtime_error("invalid alignment");
 	
 	alignment_source_target[aiter->target].push_back(aiter->source);
@@ -1125,9 +1125,9 @@ struct ExtractSCFG
     } else {
       alignment_type::const_iterator aiter_end = alignment.end();
       for (alignment_type::const_iterator aiter = alignment.begin(); aiter != aiter_end; ++ aiter) {
-	if (aiter->source >= alignment_source_target.size())
+	if (aiter->source >= static_cast<int>(alignment_source_target.size()))
 	  throw std::runtime_error("invalid alignment");
-	if (aiter->target >= alignment_target_source.size())
+	if (aiter->target >= static_cast<int>(alignment_target_source.size()))
 	  throw std::runtime_error("invalid alignment");
 	
 	alignment_source_target[aiter->source].push_back(aiter->target);
